Migrate away from flagAsRequested
