+++ RNPermissions.podspec
@@ -17,7 +17,7 @@ Pod::Spec.new do |s|
   s.source       = { :git => package["repository"]["url"], :tag => s.version }
   s.source_files = "ios/*.{h,m,mm}"
 
-  if ENV['RCT_NEW_ARCH_ENABLED'] == '1' then
+  if ENV['RCT_NEW_ARCH_ENABLED'] == "1" then
     install_modules_dependencies(s)
   else
     s.dependency   "React-Core"
