+++ ios/AppTrackingTransparency/RNPermissionHandlerAppTrackingTransparency.h
@@ -1,4 +1,4 @@
-#import "RNPermissionsHelper.h"
+#import "RNPermissionsModule.h"
 
 @interface RNPermissionHandlerAppTrackingTransparency : NSObject<RNPermissionHandler>
