+++ ios/Calendars/RNPermissionHandlerCalendars.h
@@ -1,4 +1,4 @@
-#import "RNPermissionsHelper.h"
+#import "RNPermissionsModule.h"
 
 @interface RNPermissionHandlerCalendars : NSObject<RNPermissionHandler>
