+++ ios/Contacts/RNPermissionHandlerContacts.h
@@ -1,4 +1,4 @@
-#import "RNPermissionsHelper.h"
+#import "RNPermissionsModule.h"
 
 @interface RNPermissionHandlerContacts : NSObject<RNPermissionHandler>
