+++ ios/LocationAlways/RNPermissionHandlerLocationAlways.h
@@ -1,4 +1,4 @@
-#import "RNPermissionsHelper.h"
+#import "RNPermissionsModule.h"
 
 @interface RNPermissionHandlerLocationAlways : NSObject<RNPermissionHandler>
