+++ ios/LocationWhenInUse/RNPermissionHandlerLocationWhenInUse.h
@@ -1,4 +1,4 @@
-#import "RNPermissionsHelper.h"
+#import "RNPermissionsModule.h"
 
 @interface RNPermissionHandlerLocationWhenInUse : NSObject<RNPermissionHandler>
