+++ ios/Motion/RNPermissionHandlerMotion.h
@@ -1,4 +1,4 @@
-#import "RNPermissionsHelper.h"
+#import "RNPermissionsModule.h"
 
 @interface RNPermissionHandlerMotion : NSObject<RNPermissionHandler>
