+++ ios/StoreKit/RNPermissionHandlerStoreKit.h
@@ -1,4 +1,4 @@
-#import "RNPermissionsHelper.h"
+#import "RNPermissionsModule.h"
 
 @interface RNPermissionHandlerStoreKit : NSObject<RNPermissionHandler>
